requirements.txt
streamlit
python-dotenv
openai
chromadb
azure-ai-formrecognizer
azure-core
tabulate
